Add files via upload
Updated to the code. Animation is fixed with a test function being used for updates. Needs to implement keybindings or keylisteners to use key presses to update animations instead. Also needs to implement collision logic. Added variables for borders to be used in collisions for balls and point additions. Also add a comment on what needs to be in the if for the ball to bounce of a player paddle